requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.90.0
pillow==11.2.1
pip==25.1.1
pydantic_core==2.33.2
pydantic==2.11.7
PyExifTool==0.5.6
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
watchdog==6.0.0
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
fonttools==4.58.5
